+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `createNylasSession` to `schedulingService`
+
 ## [17.0.0] - 2024-10-29
 
 ### Changed